% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createAnnoTrack.R
\name{createAnnoTrack}
\alias{createAnnoTrack}
\title{Create annotation track function}
\usage{
createAnnoTrack(TxDb = NULL, region, min.unit = 100, species = NULL,
  version = NULL)
}
\arguments{
\item{TxDb}{A TxDb object.}

\item{region}{A single GRanges object entry defining the genomic region to visualise.}

\item{min.unit}{An integer used to calculate the 'unit': width(region)/min.unit; the default is 100.}

\item{species}{[Optional] A character indicating either "Homo sapiens" or "Mus musculus". This is used only if the TxDb object is not provided.}

\item{version}{[Optional] A character specifying the annotation version (e.g., "27" or "M16"). This is used only if the TxDb object is not provided.}
}
\value{
A three-levels list containing the information to plot the track.
}
\description{
The function extracts the annotated features from the TxDb object and distributes them on different plotting layers to prevent overlaps.
}
\examples{
anno = createAnnoTrack(TxDb, region, min.unit=minimal.unit, species, version)
}
