% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTracks.R
\name{createTracks}
\alias{createTracks}
\title{Plot annotation track function}
\usage{
createTracks(x, track.names = NULL, region, use.smooth = TRUE,
  win.smooth = 1001)
}
\arguments{
\item{data}{A List object (the output of createAnnoTrack() ).}

\item{reg.range}{A vector of two integers, which define start and end for the plot left- and right-most borders.}

\item{gene.height}{A positive numeric value used as a multiplying factor for the gene height; the default is 1.}

\item{colorVar}{A character string indicating the colour of the gene lines/arrows; the default is grey.}

\item{clean}{A logical indicating whether to plot the frame around the annotation track.}
}
\value{
A ggplot object.
}
\description{
The function extracts the annotated features from the TxDb object and distributes them on different plotting layers to prevent overlaps.
}
\examples{
pl = plotanno(anno.pl, range(region.pl[, list(start, end)]))
}
