% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTxDb.R
\name{getTxDb}
\alias{getTxDb}
\title{Create TxDb object function}
\usage{
getTxDb(species, version)
}
\arguments{
\item{species}{A character indicating either "Homo sapiens" or "Mus musculus".}

\item{version}{A character specifying the annotation version (e.g., "27" or "M16").}
}
\value{
A TxDb object.
}
\description{
The function checks if local versions of Gencode annotation GFF3 and the associated Txdb exist, otherwise it dowloads the required file and create the object.
}
\examples{
TxDb = getTxDb("Homo sapiens", "27")
TxDb = getTxDb("Mus musculus", "M16")
}
